+++ VisLang/DataNodes.cs/Arithmetic/UnaryNegateOperator.cs
@@ -2,9 +2,9 @@ namespace VisLang;
 
 public class UnaryNegateOperationNodeBase : DataNode
 {
-    public float DefaultValue { get; set; } = 0f;
+    public double DefaultValue { get; set; } = 0f;
 
-    public float GetInputValue(NodeContext? context) => Inputs.FirstOrDefault()?.GetValue(context)?.AsFloat() ?? DefaultValue;
+    public double GetInputValue(NodeContext? context) => Inputs.FirstOrDefault()?.GetValue(context)?.AsFloat() ?? DefaultValue;
 
     public override Value? GetValue(NodeContext? context = null)
     {
